+++ src/stats.rs
@@ -11,20 +11,11 @@ use dactyl::{
 	NiceFloat,
 	NicePercent,
 	total_cmp,
-};
-use serde::{
-	de,
-	Deserialize,
-	Serialize,
-	ser::{
-		self,
-		SerializeStruct,
-	},
+	traits::SaturatingFrom,
 };
 use std::{
 	cmp::Ordering,
 	collections::BTreeMap,
-	fmt,
 	fs::File,
 	io::Write,
 	path::{
@@ -36,33 +27,23 @@ use std::{
 
 
 
+type HistoryData = BTreeMap<String, Stats>;
+const MAGIC: &[u8] = b"BRUNCH00";
+
+
+
 #[doc(hidden)]
 #[derive(Debug, Clone)]
 /// # History.
 ///
 /// This is triggered automatically when using the [`benches`] macro; it is
 /// not intended to be called manually.
-pub(crate) struct History(BTreeMap<String, Stats>);
+pub(crate) struct History(HistoryData);
 
 impl Default for History {
 	fn default() -> Self {
 		load_history().unwrap_or_else(|| Self(BTreeMap::default()))
 	}
-}
-
-impl<'de> Deserialize<'de> for History {
-	fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-	where D: de::Deserializer<'de> {
-		let mut out: BTreeMap<String, Stats> = de::Deserialize::deserialize(deserializer)?;
-		// Silently strip out nonsense.
-		out.retain(|k, v| ! k.is_empty() && v.is_valid());
-		Ok(Self(out))
-	}
-}
-
-impl Serialize for History {
-	fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-	where S: ser::Serializer { self.0.serialize(serializer) }
 }
 
 impl History {
@@ -79,9 +60,8 @@ impl History {
 	/// # Save.
 	pub(crate) fn save(&self) {
 		if let Some(mut f) = history_path().and_then(|f| File::create(f).ok()) {
-			if let Ok(out) = serde_json::to_vec(&self) {
-				let _res = f.write_all(&out).and_then(|_| f.flush());
-			}
+			let out = serialize(&self.0);
+			let _res = f.write_all(&out).and_then(|_| f.flush());
 		}
 	}
 }
@@ -92,10 +72,10 @@ impl History {
 /// # Runtime Stats!
 pub(crate) struct Stats {
 	/// # Total Samples.
-	total: usize,
+	total: u32,
 
 	/// # Valid Samples.
-	valid: usize,
+	valid: u32,
 
 	/// # Standard Deviation.
 	deviation: f64,
@@ -104,131 +84,10 @@ pub(crate) struct Stats {
 	mean: f64,
 }
 
-impl<'de> Deserialize<'de> for Stats {
-	fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-	where D: de::Deserializer<'de> {
-		enum Field { Total, Valid, Deviation, Mean }
-
-		impl<'de> Deserialize<'de> for Field {
-			fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-			where D: de::Deserializer<'de> {
-				struct FieldVisitor;
-
-				impl<'de> de::Visitor<'de> for FieldVisitor {
-					type Value = Field;
-
-					fn expecting(&self, f: &mut fmt::Formatter) -> fmt::Result {
-						f.write_str("`total`, `valid`, `deviation`, or `mean`")
-					}
-
-					fn visit_str<E>(self, value: &str) -> Result<Field, E>
-					where
-						E: de::Error,
-					{
-						match value {
-							"total" => Ok(Field::Total),
-							"valid" => Ok(Field::Valid),
-							"deviation" => Ok(Field::Deviation),
-							"mean" => Ok(Field::Mean),
-							_ => Err(de::Error::unknown_field(value, FIELDS)),
-						}
-					}
-				}
-
-				deserializer.deserialize_identifier(FieldVisitor)
-			}
-		}
-
-		struct StatsVisitor;
-
-		impl<'de> de::Visitor<'de> for StatsVisitor {
-			type Value = Stats;
-
-			fn expecting(&self, formatter: &mut fmt::Formatter) -> fmt::Result {
-				formatter.write_str("struct Stats")
-			}
-
-			fn visit_seq<V>(self, mut seq: V) -> Result<Stats, V::Error>
-			where V: de::SeqAccess<'de> {
-				let total: usize = seq.next_element()?
-					.ok_or_else(|| de::Error::invalid_length(0, &self))?;
-				let valid: usize = seq.next_element()?
-					.ok_or_else(|| de::Error::invalid_length(1, &self))?;
-				let deviation: f64 = seq.next_element()?
-					.ok_or_else(|| de::Error::invalid_length(2, &self))?;
-				let mean: f64 = seq.next_element()?
-					.ok_or_else(|| de::Error::invalid_length(3, &self))?;
-
-				Ok(Stats{ total, valid, deviation, mean })
-			}
-
-			fn visit_map<V>(self, mut map: V) -> Result<Stats, V::Error>
-			where V: de::MapAccess<'de> {
-				let mut total: Option<usize> = None;
-				let mut valid: Option<usize> = None;
-				let mut deviation: Option<f64> = None;
-				let mut mean: Option<f64> = None;
-
-				while let Some(key) = map.next_key()? {
-					match key {
-						Field::Total => {
-							if total.is_some() {
-								return Err(de::Error::duplicate_field("total"));
-							}
-							total.replace(map.next_value()?);
-						},
-						Field::Valid => {
-							if valid.is_some() {
-								return Err(de::Error::duplicate_field("valid"));
-							}
-							valid.replace(map.next_value()?);
-						},
-						Field::Deviation => {
-							if deviation.is_some() {
-								return Err(de::Error::duplicate_field("deviation"));
-							}
-							deviation.replace(map.next_value()?);
-						},
-						Field::Mean => {
-							if mean.is_some() {
-								return Err(de::Error::duplicate_field("mean"));
-							}
-							mean.replace(map.next_value()?);
-						},
-					}
-				}
-
-				let total = total.ok_or_else(|| de::Error::missing_field("total"))?;
-				let valid = valid.ok_or_else(|| de::Error::missing_field("valid"))?;
-				let deviation = deviation.ok_or_else(|| de::Error::missing_field("deviation"))?;
-				let mean = mean.ok_or_else(|| de::Error::missing_field("mean"))?;
-				Ok(Stats{ total, valid, deviation, mean })
-			}
-		}
-
-		const FIELDS: &[&str] = &["total", "valid", "deviation", "mean"];
-		deserializer.deserialize_struct("Stats", FIELDS, StatsVisitor)
-	}
-}
-
-impl Serialize for Stats {
-	fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-	where S: ser::Serializer {
-		let mut state = serializer.serialize_struct("Stats", 4)?;
-
-		state.serialize_field("total", &self.total)?;
-		state.serialize_field("valid", &self.valid)?;
-		state.serialize_field("deviation", &self.deviation)?;
-		state.serialize_field("mean", &self.mean)?;
-
-		state.end()
-	}
-}
-
 impl TryFrom<Vec<Duration>> for Stats {
 	type Error = BrunchError;
 	fn try_from(samples: Vec<Duration>) -> Result<Self, Self::Error> {
-		let total = samples.len();
+		let total = u32::saturating_from(samples.len());
 		if total < MIN_SAMPLES {
 			return Err(BrunchError::TooSmall(total));
 		}
@@ -237,7 +96,7 @@ impl TryFrom<Vec<Duration>> for Stats {
 		let mut calc = Abacus::from(samples);
 		calc.prune_outliers();
 
-		let valid = calc.len();
+		let valid = u32::saturating_from(calc.len());
 		if valid < MIN_SAMPLES {
 			return Err(BrunchError::TooWild);
 		}
@@ -305,7 +164,7 @@ impl Stats {
 	/// # Samples.
 	///
 	/// Return the valid/total samples.
-	pub(crate) const fn samples(self) -> (usize, usize) { (self.valid, self.total) }
+	pub(crate) const fn samples(self) -> (u32, u32) { (self.valid, self.total) }
 
 	/// # Is Valid?
 	fn is_valid(self) -> bool {
@@ -319,6 +178,74 @@ impl Stats {
 }
 
 
+
+/// # Deserialize.
+///
+/// This deserializes the inner data for a `History` object from our custom
+/// format. See `serialize` for more details.
+///
+/// This won't fail, but will strip out invalid entries as it comes across
+/// them.
+///
+/// Any time we change the version portion of our `MAGIC` constant, results
+/// from older versions will refuse to parse, resulting in an empty set.
+fn deserialize(raw: &[u8]) -> HistoryData {
+	let mut out = HistoryData::default();
+
+	// It should start with our magic header.
+	let mut raw = match raw.strip_prefix(MAGIC) {
+		Some(r) => r,
+		None => return out,
+	};
+
+	while let Some((lbl, stats, rem)) = deserialize_entry(raw) {
+		// Keep it?
+		if ! lbl.is_empty() && stats.is_valid() {
+			out.insert(lbl.to_owned(), stats);
+		}
+
+		// Are we done?
+		if rem.is_empty() { break; }
+		raw = rem;
+	}
+
+	out
+}
+
+/// # Deserialize Stat.
+///
+/// This deserializes a single benchmark entry (a label and `Stats`), returning
+/// the pieces along with the remainder of the input slice if valid.
+///
+/// This doesn't worry about key/value sanity, so will only abort if the
+/// lengths are wrong or the label cannot be stringified.
+fn deserialize_entry(raw: &[u8]) -> Option<(&str, Stats, &[u8])> {
+	const STAT_SIZE: usize = 4 + 4 + 8 + 8;
+
+	// Find the length of the label.
+	let (len, raw) = split_array::<2>(raw)?;
+	let len = u16::from_be_bytes(len) as usize;
+	if raw.len() < len + STAT_SIZE { return None; }
+
+	// Parse the label.
+	let (lbl, raw) = raw.split_at(len);
+	let lbl = std::str::from_utf8(lbl).ok()?.trim();
+
+	// Parse total, valid, deviation, and mean.
+	let (total, raw) = split_array::<4>(raw)?;
+	let total = u32::from_be_bytes(total);
+
+	let (valid, raw) = split_array::<4>(raw)?;
+	let valid = u32::from_be_bytes(valid);
+
+	let (deviation, raw) = split_array::<8>(raw)?;
+	let deviation = f64::from_be_bytes(deviation);
+
+	let (mean, raw) = split_array::<8>(raw)?;
+	let mean = f64::from_be_bytes(mean);
+
+	Some((lbl, Stats { total, valid, deviation, mean }, raw))
+}
 
 /// # History Path.
 ///
@@ -339,7 +266,57 @@ fn history_path() -> Option<PathBuf> {
 fn load_history() -> Option<History> {
 	let file = history_path()?;
 	let raw = std::fs::read(file).ok()?;
-	serde_json::from_slice(&raw).ok()
+	Some(History(deserialize(&raw)))
+}
+
+/// # Serialize.
+///
+/// This is a cheap, custom serialization structure for history. It begins with
+/// a magic header, then each entry.
+///
+/// Each entry starts with a u16 corresponding to the length of the bench name,
+/// which follows. After that, total (u32), valid (u32), deviation (f64), and
+/// mean (f64) appear.
+///
+/// All integers use Big Endian storage.
+fn serialize(history: &HistoryData) -> Vec<u8> {
+	// Magic header.
+	let mut out = MAGIC.to_vec();
+
+	// Write each section.
+	for (lbl, s) in history.iter() {
+		// We panic on long names so this won't hit, but better safe than
+		// sorry.
+		let len = match u16::try_from(lbl.len()) {
+			Ok(l) => l,
+			Err(_) => continue,
+		};
+
+		// Write the label.
+		out.extend_from_slice(&len.to_be_bytes());
+		out.extend_from_slice(lbl.as_bytes());
+
+		// Write total, valid, deviation, and mean.
+		out.extend_from_slice(&s.total.to_be_bytes());
+		out.extend_from_slice(&s.valid.to_be_bytes());
+		out.extend_from_slice(&s.deviation.to_be_bytes());
+		out.extend_from_slice(&s.mean.to_be_bytes());
+	}
+
+	out
+}
+
+/// # Split Array.
+///
+/// This takes the first `S` bytes and forms an array, then returns the rest
+/// as a slice.
+fn split_array<const S: usize>(raw: &[u8]) -> Option<([u8; S], &[u8])> {
+	if S <= raw.len() {
+		let (l, r) = raw.split_at(S);
+		let l: [u8; S] = l.try_into().ok()?;
+		Some((l, r))
+	}
+	else { None }
 }
 
 /// # Try Dir.
@@ -358,29 +335,48 @@ mod tests {
 	use super::*;
 
 	#[test]
-	fn t_stats_serde() {
-		let stat = Stats {
-			total: 2500,
-			valid: 2496,
-			deviation: 0.000000123,
-			mean: 0.0000022,
-		};
+	fn t_stats_ser() {
+		const ENTRIES: [(&str, Stats); 2] = [
+			(
+				"The First One",
+				Stats {
+					total: 2500,
+					valid: 2496,
+					deviation: 0.000000123,
+					mean: 0.0000022,
+				},
+			),
+			(
+				"The Second One",
+				Stats {
+					total: 300,
+					valid: 222,
+					deviation: 0.000400123,
+					mean: 0.0000122,
+				},
+			),
+		];
 
-		// Serialize and deserialize.
-		let s = serde_json::to_string(&stat).expect("Serialization failed.");
-		let d: Stats = serde_json::from_str(&s).expect("Deserialization failed.");
+		// Our reference.
+		let h = ENTRIES.into_iter().map(|(k, v)| (k.to_owned(), v)).collect::<HistoryData>();
 
-		// Make sure we end up where we began.
-		assert_eq!(stat.total, d.total, "Deserialization changed total.");
-		assert_eq!(stat.valid, d.valid, "Deserialization changed valid.");
-		assert!(
-			total_cmp!((stat.deviation) == (d.deviation)),
-			"Deserialization changed deviation."
-		);
-		assert!(
-			total_cmp!((stat.mean) == (d.mean)),
-			"Deserialization changed mean."
-		);
+		// Serialize it.
+		let s = serialize(&h);
+
+		// Deserialize it.
+		let d = deserialize(&s);
+
+		// The deserialized length should match our reference length.
+		assert_eq!(h.len(), d.len());
+
+		// Make sure the entries are unchanged.
+		for (lbl, stat) in ENTRIES {
+			let tmp = d.get(lbl).expect("Missing entry!");
+			assert_eq!(stat.total, tmp.total, "Total changed.");
+			assert_eq!(stat.valid, tmp.valid, "Valid changed.");
+			assert!(total_cmp!((stat.deviation) == (tmp.deviation)), "Deviation changed.");
+			assert!(total_cmp!((stat.mean) == (tmp.mean)), "Mean changed.");
+		}
 	}
 
 	#[test]